move gcc_version.h/.cpp to ansi-c
This will enable tools other than goto-cc to determine the version of gcc.
