+++ src/ansi-c/gcc_version.cpp
