+++ src/ansi-c/gcc_version.h
@@ -40,13 +40,12 @@ public:
   configt::cppt::cpp_standardt default_cxx_standard;
 
   gcc_versiont()
-    :
-    v_major(0),
-    v_minor(0),
-    v_patchlevel(0),
-    flavor(flavort::UNKNOWN),
-    default_c_standard(configt::ansi_ct::c_standardt::C89),
-    default_cxx_standard(configt::cppt::cpp_standardt::CPP98)
+    : v_major(0),
+      v_minor(0),
+      v_patchlevel(0),
+      flavor(flavort::UNKNOWN),
+      default_c_standard(configt::ansi_ct::c_standardt::C89),
+      default_cxx_standard(configt::cppt::cpp_standardt::CPP98)
   {
   }
 };
